2020/07/15 ROUTE PROTECTION

+++ controllers/videoController.js
@@ -44,8 +44,10 @@ export const postUpload = async (req, res) => {
     fileUrl: path,
     title,
     description,
+    creator: req.user.id,
   });
-
+  req.user.videos.push(newVideo);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -53,9 +55,9 @@ export const videoDetail = async (req, res) => {
   const {
     params: { id },
   } = req;
-
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator");
+    console.log(video);
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     console.log(error);
@@ -67,10 +69,13 @@ export const getEditVideo = async (req, res) => {
   const {
     params: { id },
   } = req;
-
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator.toString() !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -81,7 +86,6 @@ export const postEditVideo = async (req, res) => {
     params: { id },
     body: { title, description },
   } = req;
-
   try {
     await Video.findByIdAndUpdate({ _id: id }, { title, description });
     res.redirect(routes.videoDetail(id));
@@ -94,9 +98,13 @@ export const deleteVideo = async (req, res) => {
   const {
     params: { id },
   } = req;
-
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error;
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
